+++ TechForum/src/main/java/com/example/techforum/repository/IImageRepo.java
@@ -15,15 +15,15 @@ import java.util.List;
 
 @Repository
 @Transactional
-public interface IImageRepo extends JpaRepository<Images, Long> {
+public interface IImageRepo extends JpaRepository<Images, Integer> {
     @Modifying
     @Query(value = "insert into images (name_img,blog_id) values(?1,?2)", nativeQuery = true )
-    void createImgBlog(String nameImg, Long blog_id);
+    void createImgBlog(String nameImg, Integer blog_id);
 
     @Modifying
     @Query(value = "SELECT new com.example.techforum.dto.ImageDto(i.nameImg, i.blog) " +
             "FROM Images i JOIN Blogs b ON i.blog.id = b.id " +
             "WHERE b.id = :id")
-    List<ImageDto> findImageDtosByBlogId(@Param("id") Long blogId);
+    List<ImageDto> findImageDtosByBlogId(@Param("id") Integer blogId);
 
 }
