+++ TechForum/src/main/java/com/example/techforum/service/blogstorage/IBlogStorageService.java
@@ -9,9 +9,9 @@ import org.springframework.data.domain.Pageable;
 import java.util.List;
 
 public interface IBlogStorageService {
-    Page<BlogStorage> findAllByUserId(Long id,Pageable pageable);
-    String toggleBookMark(Long blogId, Long userId);
-    void delete(Long id);
-    int getCount(Long blogId);
-    Boolean isBlogBookMarkedByUser(Long blogId, Long userId);
+    Page<BlogStorage> findAllByUserId(Integer id,Pageable pageable);
+    String toggleBookMark(Integer blogId, Integer userId);
+    void delete(Integer id);
+    int getCount(Integer blogId);
+    Boolean isBlogBookMarkedByUser(Integer blogId, Integer userId);
 }
