+++ TechForum/src/main/java/com/example/techforum/service/views/IViewsService.java
@@ -1,6 +1,6 @@
 package com.example.techforum.service.views;
 
 public interface IViewsService {
-    String toggleView(Long blogId, Long userId);
-    int getViewCount(Long blogId);
+    String toggleView(Integer blogId, Integer userId);
+    int getViewCount(Integer blogId);
 }
